// src/app/models/student.model.ts
import { Gender } from './gender.enum';
export class Student {

    date?: Date;
    studentId?: any;
    studentName ?: string;
    studentNrc?: string;
    age?: number;
    dateOfBirth?: Date;
    fatherName?: string;    
    gender?: Gender;
    address?: string;
    township?: string;
    photo?: string;
}